requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.30.1
prometheus-client==0.20.0
PyYAML==6.0.1
psutil==5.9.8
gunicorn==21.2.0
